Bug 53039 - HTTP Request : Be able to handle responses which size exceeds 2147483647 bytes Bugzilla Id: 53039
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1764397 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: caaf9e666c4d0facd5c95deebaed5d27911666fc

+++ src/core/org/apache/jmeter/report/core/Sample.java
@@ -260,8 +260,8 @@ public class Sample {
      *
      * @return the number of received bytes stored in the sample
      */
-    public int getReceivedBytes() {
-        return getData(int.class, CSVSaveService.CSV_BYTES).intValue();
+    public long getReceivedBytes() {
+        return getData(long.class, CSVSaveService.CSV_BYTES).longValue();
     }
 
     /**
